Fix including last segment in file
When a file doesn't include a silence at the end, the last segment
wasn't being included. Always include the segment from the last silence
to the end. This should be fine even if the last silence is, in fact, at
the end of the file, because it just means we include a little more
silence in the output.